Restructure according to experiments
